tracking of free memory in allocator instead of header
this may lose the padding bytes when deallocating a region that was shifted up
due to alignment requirements. if a new region is allocated right after,
that has a weaker alignment, it will still be aligned at the stricter
address. but it allows saving 64 bytes in each header (coming up) which
should be worth it.